+++ src/cohort_simulator.py
@@ -139,10 +139,10 @@ def simulate_cohorts_SI(
         pi = 0
 
     # quartiles = np.array([1, 0.75, 0.5, 0.25, 0])
-    Phi_parti = np.ones((Nt))
-    invest_mat = np.ones((Nt, Ntype, Nc), dtype=int)
-    parti_wealth_group = np.ones((Nt, 4))
-    parti_age_group = np.ones((Nt, 4))
+    Phi_parti = np.ones((Nt), dtype=np.float16)
+    invest_mat = np.ones((Nt, Ntype, Nc), dtype=np.int8)
+    parti_wealth_group = np.ones((Nt, 4), dtype=np.float16)
+    parti_age_group = np.ones((Nt, 4), dtype=np.float16)
 
     # if mode_trade == 'partial_constraint_rich' or mode_trade == 'partial_constraint_old':
     #     popu_can_short= np.zeros((Nt))
@@ -162,9 +162,9 @@ def simulate_cohorts_SI(
     mu_S = np.zeros(Nt)
     sigma_S = np.zeros(Nt)
     beta = np.zeros(Nt)
-    Delta_bar_parti = np.zeros((Nt))  # consumption weighted estimation error of the stock market participants
-    Delta_tilde_parti = np.zeros((Nt))  # wealth weighted estimation error of the stock market participants
-    parti = np.ones((Nt))  # participation rate
+    Delta_bar_parti = np.zeros((Nt), dtype=np.float16)  # consumption weighted estimation error of the stock market participants
+    Delta_tilde_parti = np.zeros((Nt), dtype=np.float16)  # wealth weighted estimation error of the stock market participants
+    parti = np.ones((Nt), dtype=np.float16)  # participation rate
 
     # upperbound = np.arange(10,55,5)
     # theta_t_matrix = np.zeros((Nt, len(upperbound)))
